bist-analiz-new/bist-analiz-new/requirements.txt
streamlit
pandas
plotly
yfinance
python-dotenv
langchain